Undo rename of enums package.
