+++ stf-desktop-automation-lib/src/main/java/com/github/jeansantos38/stf/enums/UiAction.java
