+++ stf-http-client-lib/src/main/java/com/github/jeansantos38/stf/enums/http/HttpRequestLogLevel.java
@@ -1,4 +1,4 @@
-package com.github.jeansantos38.stf.enums.http;
+package com.github.jeansantos38.stf.enums.ui.http;
 
 /************************************************************
  *  © Copyright 2019 HP Development Company, L.P.
