+++ stf-misc-lib/src/main/java/com/github/jeansantos38/stf/enums/wait/ThreadWait.java
@@ -1,4 +1,4 @@
-package com.github.jeansantos38.stf.enums.wait;
+package com.github.jeansantos38.stf.enums.ui.wait;
 
 /************************************************************
  *  © Copyright 2019 HP Development Company, L.P.
