+++ stf-webdriver-lib/src/main/java/com/github/jeansantos38/stf/enums/webdriver/SelectorType.java
@@ -1,4 +1,4 @@
-package com.github.jeansantos38.stf.enums.serialization;
+package com.github.jeansantos38.stf.enums.ui.webdriver;
 
 /************************************************************
  *  © Copyright 2019 HP Development Company, L.P.
@@ -6,6 +6,6 @@ package com.github.jeansantos38.stf.enums.serialization;
  *
  *  Smart Test Framework
  ************************************************************/
-public enum SerializationType {
-    JSON, XML, YAML
+public enum SelectorType {
+    BY_ID, BY_XPATH, BY_CLASS_NAME, BY_CSS, BY_NAME, BY_TAG_NAME, BY_LINK, BY_PARTIAL_LINK
 }
