+++ stf-webdriver-lib/src/main/java/com/github/jeansantos38/stf/enums/webdriver/WebDriverAction.java
@@ -1,4 +1,4 @@
-package com.github.jeansantos38.stf.enums.webdriver;
+package com.github.jeansantos38.stf.enums.ui.webdriver;
 
 /************************************************************
  *  © Copyright 2019 HP Development Company, L.P.
